Track degrees of edges
When creating a new Pair, just use as the new degree, the sum of the two old
degrees (or one for a Pair of a new tetrahedron). However, add one more to the
degree value if the two faces being joined belong to the same tetrahedron. This
way, the only way to get degree (of the Pair) ≤ 3 is if the corresponding edge
in the triangulation either has degree < 3, or it has degree = 3 and also
belongs to 3 distinct tetrahedra, both of which mean we have a bad
configuration.